Add new control event STATUS_WORK to signal to controller current status.
Since on Android it is not possible to hold a wake lock in purely native code,
we introduce a new control event to signal the controller to hold a wake lock
for us, and the STATUS_WORK value indicates the current state (true if we are in
the middle of uninterruptable work, false if not).

This is very finicky as we are trying to force a buffer flush rather than rely
on the existing queue as the event needs to arrive at the controller as soon as
possible.